+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 ### Changed
 - Fix missing activity param import (#147)
 - Remove page.screen plugin (#150)
+- Remove ContributesInjector option from Page template (#167)
 
 ## [2.1.5] - 2025-10-30
 ### Changed